requirements.txt
torch
torchvision
torchaudio
matplotlib
seaborn
scikit-learn
numpy
pillow
opencv-python
gradio
tqdm
